+++ src/plot/layer/geom/area.rs
@@ -60,10 +60,6 @@ impl GeomTrait for Area {
         Some(&["pos1"])
     }
 
-    fn needs_stat_transform(&self, _aesthetics: &Mappings) -> bool {
-        true
-    }
-
     fn apply_stat_transform(
         &self,
         query: &str,