+++ src/plot/layer/geom/ribbon.rs
@@ -47,10 +47,6 @@ impl GeomTrait for Ribbon {
         Some(&["pos1"])
     }
 
-    fn needs_stat_transform(&self, _aesthetics: &Mappings) -> bool {
-        true
-    }
-
     fn apply_stat_transform(
         &self,
         query: &str,